+++ h3ron/src/collections/partitioned/serde.rs
@@ -1,5 +1,5 @@
 use std::fmt;
-use std::hash::{BuildHasher, Hash};
+use std::hash::Hash;
 use std::iter::FromIterator;
 use std::marker::PhantomData;
 
@@ -7,13 +7,12 @@ use serde::de::{MapAccess, Visitor};
 use serde::ser::SerializeMap;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
-use crate::collections::{RandomState, ThreadPartitionedMap};
+use crate::collections::ThreadPartitionedMap;
 
-impl<K, V, S> Serialize for ThreadPartitionedMap<K, V, S>
+impl<K, V, const N: usize> Serialize for ThreadPartitionedMap<K, V, N>
 where
     K: Hash + Eq + Send + Sync + Serialize,
     V: Send + Sync + Serialize,
-    S: BuildHasher + Default + Send + Clone,
 {
     fn serialize<SER>(&self, serializer: SER) -> Result<SER::Ok, SER::Error>
     where
@@ -29,17 +28,17 @@ where
     }
 }
 
-struct ThreadPartitionedMapVisitor<K, V> {
+struct ThreadPartitionedMapVisitor<K, V, const N: usize> {
     #[allow(clippy::type_complexity)]
-    marker: PhantomData<fn() -> ThreadPartitionedMap<K, V, RandomState>>,
+    marker: PhantomData<fn() -> ThreadPartitionedMap<K, V, N>>,
 }
 
-impl<'de, K, V> Visitor<'de> for ThreadPartitionedMapVisitor<K, V>
+impl<'de, K, V, const N: usize> Visitor<'de> for ThreadPartitionedMapVisitor<K, V, N>
 where
     K: Hash + Eq + Send + Sync + Deserialize<'de>,
     V: Send + Sync + Deserialize<'de>,
 {
-    type Value = ThreadPartitionedMap<K, V, RandomState>;
+    type Value = ThreadPartitionedMap<K, V, N>;
 
     fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
         formatter.write_str("ThreadPartitionedMap failed")
@@ -57,7 +56,7 @@ where
     }
 }
 
-impl<'de, K, V> Deserialize<'de> for ThreadPartitionedMap<K, V, RandomState>
+impl<'de, K, V, const N: usize> Deserialize<'de> for ThreadPartitionedMap<K, V, N>
 where
     K: Hash + Eq + Send + Sync + Deserialize<'de>,
     V: Send + Sync + Deserialize<'de>,
@@ -81,12 +80,12 @@ mod tests {
     #[test]
     fn serde_roundtrip() {
         let in_vec: Vec<_> = (0_u64..1_000).map(|i| (i, i)).collect();
-        let tpm = ThreadPartitionedMap::from_iter(in_vec.clone());
+        let tpm: ThreadPartitionedMap<_, _, 6> = ThreadPartitionedMap::from_iter(in_vec.clone());
 
         let byte_data = bincode::serialize(&tpm).unwrap();
 
         let mut tpm_de =
-            bincode::deserialize::<ThreadPartitionedMap<u64, u64>>(&byte_data).unwrap();
+            bincode::deserialize::<ThreadPartitionedMap<u64, u64, 6>>(&byte_data).unwrap();
 
         assert_eq!(tpm_de.len(), tpm.len());
         let mut out_vec: Vec<_> = tpm_de.drain().collect();